requirements.txt
numpy
polyline
requests
selenium
Pillow